[BUG-0000][RB-6126]
This is the 6th drop in a series of changes, refactoring, etc to come up with a DSL/script for Cython P5 NFS Test Case.

This is still an experimental script & is expected to have code churns till it matures to beta stage.

This is what we have achieved so far:

- A DSL for scripting
- Core consists of around ~700 loc (including comments & line spaces).
- No debug, no logging. The automation/core will tell the issues proactively.
- Execute remote commands
- Run the output over a chain of operators
- Store specific values as the chain of operators gets executed
- Use specific values as the chain of commands get processed
- Conditional execution of remote commands
  i.e. execute based on outputs of previous commands
- DSL specs:
  - ssh (multiple variants)
  - to
  - run
  - storeAs
  - then (multiple variants)
  - can use groovy operators in above specs




git-svn-id: svn+ssh://10.10.1.102/svnrepo/devops@43 20142372-ea64-11e5-9694-1dd7e65929dd